Update project folder structure
