--- src/frontend/ts/src/common.ts
@@ -1,421 +1,0 @@
-export enum ExtractorErrorCode {
-  MissingExport,
-  AppMissingDefaultExport,
-  UnknownType,
-  MultipleGenericType,
-  InvalidDataSourceDefinition,
-  InvalidPartialType,
-  InvalidIncludeTree,
-  InvalidAttributeModifier,
-  InvalidApiMethodModifier,
-  UnknownNavigationPropertyReference,
-  InvalidNavigationPropertyReference,
-  MissingNavigationPropertyReference,
-  MissingManyToManyUniqueId,
-  MissingPrimaryKey,
-  MissingDatabaseBinding,
-  MissingWranglerEnv,
-  TooManyWranglerEnvs,
-  MissingFile,
-}
-
-const errorInfoMap: Record<
-  ExtractorErrorCode,
-  { description: string; suggestion: string }
-> = {
-  [ExtractorErrorCode.MissingExport]: {
-    description: "All Cloesce types must be exported.",
-    suggestion: "Add `export` to the class definition.",
-  },
-  [ExtractorErrorCode.AppMissingDefaultExport]: {
-    description: "app.cloesce.ts does not export a CloesceApp by default",
-    suggestion: "Export an instantiated CloesceApp in app.cloesce.ts",
-  },
-  [ExtractorErrorCode.UnknownType]: {
-    description: "Encountered an unknown or unsupported type",
-    suggestion: "Refer to the documentation on valid Cloesce TS types",
-  },
-  [ExtractorErrorCode.InvalidPartialType]: {
-    description: "Partial types must only contain a model or plain old object",
-    suggestion: "Refer to the documentation on valid Cloesce TS types",
-  },
-  [ExtractorErrorCode.MultipleGenericType]: {
-    description: "Cloesce does not yet support types with multiple generics",
-    suggestion:
-      "Simplify your type to use only a single generic parameter, ie Foo<T>",
-  },
-  [ExtractorErrorCode.InvalidDataSourceDefinition]: {
-    description:
-      "Data Sources must be explicitly typed as a static Include Tree",
-    suggestion:
-      "Declare your data source as `static readonly _: IncludeTree<Model>`",
-  },
-  [ExtractorErrorCode.InvalidIncludeTree]: {
-    description: "Invalid Include Tree",
-    suggestion:
-      "Include trees must only contain references to a model's navigation properties.",
-  },
-  [ExtractorErrorCode.InvalidAttributeModifier]: {
-    description:
-      "Attributes can only be public on a Model, Plain Old Object or Wrangler Environment",
-    suggestion: "Change the attribute modifier to just `public`",
-  },
-  [ExtractorErrorCode.InvalidApiMethodModifier]: {
-    description:
-      "Model methods must be public if they are decorated as GET, POST, PUT, PATCH",
-    suggestion: "Change the method modifier to just `public`",
-  },
-  [ExtractorErrorCode.UnknownNavigationPropertyReference]: {
-    description: "Unknown Navigation Property Reference",
-    suggestion:
-      "Verify that the navigation property reference model exists, or create a model.",
-  },
-  [ExtractorErrorCode.InvalidNavigationPropertyReference]: {
-    description: "Invalid Navigation Property Reference",
-    suggestion: "Ensure the navigation property points to a valid model field",
-  },
-  [ExtractorErrorCode.MissingNavigationPropertyReference]: {
-    description: "Missing Navigation Property Reference",
-    suggestion:
-      "Navigation properties require a foreign key model attribute reference",
-  },
-  [ExtractorErrorCode.MissingManyToManyUniqueId]: {
-    description: "Missing unique id on Many to Many navigation property",
-    suggestion:
-      "Define a unique identifier field for the Many-to-Many relationship",
-  },
-  [ExtractorErrorCode.MissingPrimaryKey]: {
-    description: "Missing primary key on a model",
-    suggestion: "Add a primary key field to your model (e.g., `id: number`)",
-  },
-  [ExtractorErrorCode.MissingDatabaseBinding]: {
-    description: "Missing a database binding in the WranglerEnv definition",
-    suggestion: "Add a `D1Database` to your WranglerEnv",
-  },
-  [ExtractorErrorCode.MissingWranglerEnv]: {
-    description: "Missing a wrangler environment definition in the project",
-    suggestion: "Add a @WranglerEnv class in your project.",
-  },
-  [ExtractorErrorCode.TooManyWranglerEnvs]: {
-    description: "Too many wrangler environments defined in the project",
-    suggestion: "Consolidate or remove unused @WranglerEnv's",
-  },
-  [ExtractorErrorCode.MissingFile]: {
-    description: "A specified input file could not be found",
-    suggestion: "Verify the input file path is correct",
-  },
-};
-
-export function getErrorInfo(code: ExtractorErrorCode) {
-  return errorInfoMap[code];
-}
-
-export class ExtractorError {
-  context?: string;
-  snippet?: string;
-
-  constructor(public code: ExtractorErrorCode) {}
-
-  addContext(fn: (val: string | undefined) => string | undefined) {
-    this.context = fn(this.context ?? "");
-  }
-}
-
-type DeepPartialInner<T> = T extends (infer U)[]
-  ? DeepPartialInner<U>[]
-  : T extends object
-    ? { [K in keyof T]?: DeepPartialInner<T[K]> }
-    : T | (null extends T ? null : never);
-
-/**
- * Recursively makes all properties of a type optional — including nested objects and arrays.
- *
- * Similar to TypeScript's built-in `Partial<T>`, but applies the transformation deeply across
- * all nested structures. Useful for defining "patch" or "update" objects where only a subset
- * of properties may be provided.
- *
- * **Apart of the Cloesce method grammar**, meaning the type can be apart of method parameters
- * or return types and the generated workers and client API will act accordingly.
- *
- * @template T
- * The target type to make deeply partial.
- *
- * @remarks
- * - **Objects:** All properties become optional, and their values are recursively wrapped in `DeepPartial`.
- * - **Arrays:** Arrays are preserved, but their elements are recursively made partial.
- * - **Scalars:** Primitive values (string, number, boolean, etc.) remain unchanged.
- * - **Nullable types:** If `null` is assignable to the type, it remains allowed.
- *
- * @example
- * ```ts
- * class User {
- *   id: string;
- *   profile: {
- *     name: string;
- *     age: number;
- *   };
- *   tags: string[];
- * }
- *
- * // The resulting type:
- * // {
- * //   id?: string;
- * //   profile?: { name?: string; age?: number };
- * //   tags?: (string | undefined)[];
- * // }
- * type PartialUser = DeepPartial<User>;
- *
- * const patch: PartialUser = {
- *   profile: { age: 30 } // ok
- * };
- * ```
- */
-export type DeepPartial<T> = DeepPartialInner<T> & { __brand?: "Partial" };
-
-export class Either<L, R> {
-  private constructor(
-    private readonly inner: { ok: true; right: R } | { ok: false; left: L },
-  ) {}
-
-  get value(): L | R {
-    return this.inner.ok ? this.inner.right : this.inner.left;
-  }
-
-  static left<L, R = never>(value: L): Either<L, R> {
-    return new Either({ ok: false, left: value });
-  }
-
-  static right<R, L = never>(value: R): Either<L, R> {
-    return new Either({ ok: true, right: value });
-  }
-
-  isLeft(): this is Either<L, never> {
-    return !this.inner.ok;
-  }
-
-  isRight(): this is Either<never, R> {
-    return this.inner.ok;
-  }
-
-  unwrap(): R {
-    if (!this.inner.ok) {
-      throw new Error("Tried to unwrap a Left value");
-    }
-    return this.inner.right;
-  }
-
-  unwrapLeft(): L {
-    if (this.inner.ok) {
-      throw new Error("Tried to unwrapLeft a Right value");
-    }
-    return this.inner.left;
-  }
-
-  map<B>(fn: (val: R) => B): Either<L, B> {
-    return this.inner.ok
-      ? Either.right(fn(this.inner.right))
-      : Either.left(this.inner.left);
-  }
-
-  mapLeft<B>(fn: (val: L) => B): Either<B, R> {
-    return this.inner.ok
-      ? Either.right(this.inner.right)
-      : Either.left(fn(this.inner.left));
-  }
-}
-
-export class HttpResult<T = unknown> {
-  public constructor(
-    public ok: boolean,
-    public status: number,
-    public headers: Headers,
-    public data?: T,
-    public message?: string,
-  ) {}
-
-  static ok<T>(status: number, data?: T, init?: HeadersInit): HttpResult {
-    const headers: Headers = new Headers(
-      init ?? {
-        "Content-Type": "application/json",
-      },
-    );
-
-    return new HttpResult<T>(true, status, headers, data, undefined);
-  }
-
-  static fail(status: number, message?: string, init?: HeadersInit) {
-    const headers: Headers = new Headers(
-      init ?? {
-        "Content-Type": "application/json",
-      },
-    );
-
-    return new HttpResult<never>(false, status, headers, undefined, message);
-  }
-
-  toResponse(): Response {
-    const body = () => {
-      if (this.status === 204) {
-        return undefined;
-      }
-
-      return this.ok
-        ? JSON.stringify({
-            data: this.data,
-          })
-        : this.message;
-    };
-
-    return new Response(body(), {
-      status: this.status,
-      headers: this.headers,
-    });
-  }
-
-  static async fromResponse(
-    response: Response,
-    ctor?: any,
-    array: boolean = false,
-  ): Promise<HttpResult<any>> {
-    if (response.status < 400) {
-      const json = await response.json();
-
-      // The `fromJson` method is generated on every single model.
-      let data = json.data;
-      if (ctor) {
-        if (array) {
-          for (let i = 0; i < data.length; i++) {
-            data[i] = ctor.fromJson(data[i]);
-          }
-        } else {
-          data = ctor.fromJson(data);
-        }
-      }
-
-      return new HttpResult(true, response.status, response.headers, data);
-    }
-
-    return new HttpResult(
-      false,
-      response.status,
-      response.headers,
-      undefined,
-      await response.text(),
-    );
-  }
-}
-
-export type KeysOfType<T, U> = {
-  [K in keyof T]: T[K] extends U ? (K extends string ? K : never) : never;
-}[keyof T];
-
-export type CrudKind = "SAVE" | "GET" | "LIST";
-
-export type CidlType =
-  | "Void"
-  | "Integer"
-  | "Real"
-  | "Text"
-  | "Blob"
-  | "DateIso"
-  | "Boolean"
-  | { DataSource: string }
-  | { Inject: string }
-  | { Object: string }
-  | { Partial: string }
-  | { Nullable: CidlType }
-  | { Array: CidlType }
-  | { HttpResult: CidlType };
-
-export function isNullableType(ty: CidlType): boolean {
-  return typeof ty === "object" && ty !== null && "Nullable" in ty;
-}
-
-export enum HttpVerb {
-  GET = "GET",
-  POST = "POST",
-  PUT = "PUT",
-  PATCH = "PATCH",
-  DELETE = "DELETE",
-}
-
-export interface NamedTypedValue {
-  name: string;
-  cidl_type: CidlType;
-}
-
-export interface ModelAttribute {
-  value: NamedTypedValue;
-  foreign_key_reference: string | null;
-}
-
-export interface ModelMethod {
-  name: string;
-  is_static: boolean;
-  http_verb: HttpVerb;
-  return_type: CidlType | null;
-  parameters: NamedTypedValue[];
-}
-
-export type NavigationPropertyKind =
-  | { OneToOne: { reference: string } }
-  | { OneToMany: { reference: string } }
-  | { ManyToMany: { unique_id: string } };
-
-export interface NavigationProperty {
-  var_name: string;
-  model_name: string;
-  kind: NavigationPropertyKind;
-}
-
-export function getNavigationPropertyCidlType(
-  nav: NavigationProperty,
-): CidlType {
-  return "OneToOne" in nav.kind
-    ? { Object: nav.model_name }
-    : { Array: { Object: nav.model_name } };
-}
-
-export interface Model {
-  name: string;
-  primary_key: NamedTypedValue;
-  attributes: ModelAttribute[];
-  navigation_properties: NavigationProperty[];
-  methods: Record<string, ModelMethod>;
-  data_sources: Record<string, DataSource>;
-  cruds: CrudKind[];
-  source_path: string;
-}
-
-export interface PlainOldObject {
-  name: string;
-  attributes: NamedTypedValue[];
-  source_path: string;
-}
-
-export interface CidlIncludeTree {
-  [key: string]: CidlIncludeTree;
-}
-
-export const NO_DATA_SOURCE = "none";
-export interface DataSource {
-  name: string;
-  tree: CidlIncludeTree;
-}
-
-export interface WranglerEnv {
-  name: string;
-  source_path: string;
-  db_binding: string;
-  vars: Record<string, CidlType>;
-}
-
-export interface CloesceAst {
-  version: string;
-  project_name: string;
-  language: string;
-  wrangler_env: WranglerEnv;
-  models: Record<string, Model>;
-  poos: Record<string, PlainOldObject>;
-  app_source: string | null;
-}
