+++ src/frontend/ts/src/router/crud.ts
@@ -1,6 +1,7 @@
 import { D1Database } from "@cloudflare/workers-types/experimental";
-import { HttpResult, NO_DATA_SOURCE } from "../common.js";
 import { IncludeTree, Orm } from "../ui/backend.js";
+import { HttpResult } from "../ui/common.js";
+import { NO_DATA_SOURCE } from "../ast.js";
 
 /**
  * A wrapper for Model Instances, containing definitions for built-in CRUD methods.
