+++ src/frontend/ts/src/ui/client.ts
@@ -1,2 +1,5 @@
-export type { DeepPartial } from "../common.js";
-export { HttpResult, Either } from "../common.js";
+/**
+ * Types accessible via cloesce/client
+ */
+export type { DeepPartial } from "./common.js";
+export { HttpResult, Either } from "./common.js";
