+++ src/frontend/ts/tsconfig.json
@@ -11,7 +11,8 @@
     "resolveJsonModule": true,
     "declaration": true,
     "declarationMap": true,
-    "emitDeclarationOnly": false
+    "emitDeclarationOnly": false,
+    "stripInternal": true
   },
   "include": ["src"]
 }
