+++ tests/fixtures/regression/bools_dates_ints/client.ts
@@ -16,7 +16,7 @@ export class Weather {
     baseUrl.searchParams.append('id', String(id));
     baseUrl.searchParams.append('__datasource', String(__datasource));
     const res = await fetchImpl(baseUrl, { method: "GET" });
-    return await HttpResult.fromResponse(res, Weather, false);
+    return await HttpResult.fromResponse<Weather>(res, Weather, false);
   }
   static async save(
         model: DeepPartial<Weather>,
@@ -32,7 +32,7 @@ export class Weather {
             __datasource
       })
     });
-    return await HttpResult.fromResponse(res, Weather, false);
+    return await HttpResult.fromResponse<Weather>(res, Weather, false);
   }
 
   static fromJson(data: any): Weather {
