+++ tests/fixtures/regression/middleware/client.ts
@@ -18,14 +18,14 @@ export class Model {
   ): Promise<HttpResult<void>> {
     const baseUrl = new URL(`http://localhost:5002/api/Model/blockedMethod`);
     const res = await fetchImpl(baseUrl, { method: "GET" });
-    return await HttpResult.fromResponse(res);
+    return await HttpResult.fromResponse<void>(res);
   }
   static async getInjectedThing(
     fetchImpl: typeof fetch = fetch
   ): Promise<HttpResult<InjectedThing>> {
     const baseUrl = new URL(`http://localhost:5002/api/Model/getInjectedThing`);
     const res = await fetchImpl(baseUrl, { method: "GET" });
-    return await HttpResult.fromResponse(res, InjectedThing, false);
+    return await HttpResult.fromResponse<InjectedThing>(res, InjectedThing, false);
   }
   static async save(
         model: DeepPartial<Model>,
@@ -41,7 +41,7 @@ export class Model {
             __datasource
       })
     });
-    return await HttpResult.fromResponse(res, Model, false);
+    return await HttpResult.fromResponse<Model>(res, Model, false);
   }
 
   static fromJson(data: any): Model {
